% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf.R
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{p2r_sa_component}
\alias{p2r_sa_decomposition}
\alias{p2r_sa_diagnostics}
\alias{p2r_spec_benchmarking}
\alias{r2p_spec_benchmarking}
\title{Java Utility Functions}
\usage{
p2r_sa_component(p)

p2r_sa_decomposition(p, full = F)

p2r_sa_diagnostics(p)

p2r_spec_benchmarking(p)

r2p_spec_benchmarking(r)
}
\arguments{
\item{p, r, full}{parameters.}
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
