% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R, R/generics.R
\name{sadecomposition}
\alias{sadecomposition}
\alias{print.JD3_SADECOMPOSITION}
\alias{sa.decomposition}
\title{Generic Function for Seasonal Adjustment Decomposition}
\usage{
sadecomposition(y, sa, t, s, i, mul)

\method{print}{JD3_SADECOMPOSITION}(x, n_last_obs = frequency(x$series), ...)

sa.decomposition(x, ...)
}
\arguments{
\item{y, sa, t, s, i, mul}{seasonal adjustment decomposition parameters.}

\item{x}{the object to print.}

\item{n_last_obs}{number of observations to print (by default equal to the frequency of the series).}

\item{...}{further arguments (ignored).}
}
\value{
\code{"JD3_SADECOMPOSITION"} object.
}
\description{
Generic function to format the seasonal adjustment decomposition components.
\code{sa.decomposition()} is a generic function defined in other packages.
}
