% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality.canovahansen}
\alias{seasonality.canovahansen}
\title{Title}
\usage{
seasonality.canovahansen(data, p0, p1, np, original = FALSE)
}
\arguments{
\item{p0}{Initial periodicity (included)}

\item{p1}{Final periodicity (included)}

\item{np}{Number of periodicities equally spaced in [p0,p1]}

\item{original}{True for original algorithm, False for solution proposed by T. Proietti (based on Ox code)}
}
\value{

}
\description{
Title
}
