% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality.friedman}
\alias{seasonality.friedman}
\title{Friedman Seasonality Test}
\usage{
seasonality.friedman(data, period, nyears = 0)
}
\arguments{
\item{data}{the input time series.}

\item{period}{the periodicity of the data.}

\item{nyears}{\code{integer} that corresponds to number of periods number of periods starting from the end of the series:
in periods (positive value) or years (negative values).
By default (\code{nyears = 0}), the entire sample is used.}
}
\value{

}
\description{
Friedman Seasonality Test
}
\details{
Non parametric test ("ANOVA"-type)
}
