% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality.qs}
\alias{seasonality.qs}
\title{QS (modified seasonal Ljung-Box) test}
\usage{
seasonality.qs(data, period, nyears = 0)
}
\arguments{
\item{data}{Data (numerics)}

\item{period}{Tested periodicity}

\item{nyears}{Number of cycles considered in the test, at the end of the series. 0 for using the whole series}
}
\value{

}
\description{
QS (modified seasonal Ljung-Box) test
}
\examples{
seasonality.qs(rjd3toolkit::ABS$X0.2.09.10.M, 12)
seasonality.qs(rjd3toolkit::randomsT(2, 1000), 7)
}
