% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality.combined}
\alias{seasonality.combined}
\title{“X12” Test On Seasonality}
\usage{
seasonality.combined(data, period, firstperiod = 1, mul = T)
}
\arguments{
\item{data}{the input time series.}

\item{period}{the periodicity of the data.}

\item{mul}{boolean indicating if the seasonal decomposition is multiplicative (\code{mul = TRUE}) or additive (\code{mul = FALSE}).}
}
\value{

}
\description{
“X12” Test On Seasonality
}
\details{
Combined test on the presence of identifiable seasonality (see Ladiray and Quenneville, 1999).
}
