% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality.kruskalwallis}
\alias{seasonality.kruskalwallis}
\title{Kruskall-Wallis Seasonality Test}
\usage{
seasonality.kruskalwallis(data, period, nyears = 0)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.}

\item{nyears}{Number of number of periods number of cycles considered in the test, at the end of the series:
in periods (positive value) or years (negative values).
By default (\code{nyears = 0}), the entire sample is used.}
}
\value{

}
\description{
Kruskall-Wallis Seasonality Test
}
\details{
Non parametric test on the ranks
}
