% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{seasonality.qs}
\alias{seasonality.qs}
\title{QS Seasonality Test}
\usage{
seasonality.qs(data, period, nyears = 0)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.}

\item{nyears}{Number of number of periods number of cycles considered in the test, at the end of the series:
in periods (positive value) or years (negative values).
By default (\code{nyears = 0}), the entire sample is used.}
}
\value{
A \code{c("JD3_TEST", "JD3")} object (see \code{\link[rjd3toolkit]{statisticaltest}} for details).
}
\description{
QS (modified seasonal Ljung-Box) test
}
\examples{
seasonality.qs(rjd3toolkit::ABS$X0.2.09.10.M, 12)
seasonality.qs(rjd3toolkit::randomsT(2, 1000), 7)
}
